Create a profiler-solution and profiler folder for the exercise
