+++ pom.xml
@@ -16,6 +16,7 @@
 
     <modules>
         <module>tested-app</module>
+        <module>profiler-solution</module>
         <module>profiler</module>
     </modules>
