--- profiler/src/main/java/irach/demo/profiler/LogProfiler.java
@@ -1,18 +1,0 @@
-package irach.demo.profiler;
-
-public class LogProfiler
-{
-  public static final LogProfiler $_INSTANCE = new LogProfiler();
-
-  int count = 0;
-
-  private LogProfiler()
-  {
-  }
-
-  public void logMethodCalled()
-  {
-    count++;
-    System.out.println("[Agent] log method was called " + count + " times");
-  }
-}
